Make function std::ping::ping concurrency safe
The function `ping` so far did not check the ident received. As a consequence
the function `ping` might returned wrong results when used simultaneously from
different threads, see https://github.com/aisk/ping/issues/6 for more details.

This commit fixes the issue by checking the ident of the reply and comparing it
with the expected ident.

Changes:
* Changed type of `timeout` from `Option<Duration>` to `Duration`. This avoids
  an additional error check in new code.
* Receive echo in a loop until either the correct ident was received or a
  timeout occured.